chore(format, C++ warnings): reformat and eliminate C++ warnings.
Fixed all minutiae like unsigned comparing to signed problem, all -Wall and -Wextra warnings are fixed.

Complete the README file.

+++ resnet/include/functional/add.hpp
@@ -2,11 +2,19 @@
 */#ifndef TENSORCORE_RESNET_COMMON_FUNCTIONAL_ADD_HPP
 #define TENSORCORE_RESNET_COMMON_FUNCTIONAL_ADD_HPP
 
-#include "common.h"
+#include "common.hpp"
 
-void add_(float *Result, const float *adder, int length, Impl::DeviceType device_type, cudaStream_t stream = cudaStream_t(nullptr));
-[[maybe_unused]] void add(float *Result, const float *adder_a, const float *adder_b, int length, Impl::DeviceType device_type);
-void relu_(float *Result, int length, Impl::DeviceType device_type);
+void add_(float *Result,
+          const float *adder,
+          size_t length,
+          Impl::DeviceType device_type,
+          cudaStream_t stream = cudaStream_t(nullptr));
+[[maybe_unused]] void add(float *Result,
+                          const float *adder_a,
+                          const float *adder_b,
+                          int length,
+                          Impl::DeviceType device_type);
+void relu_(float *Result, size_t length, Impl::DeviceType device_type);
 // Fuse
-void add_relu_(float *Result, const float *adder, int length, Impl::DeviceType device_type);
+void add_relu_(float *Result, const float *adder, size_t length, Impl::DeviceType device_type);
 #endif //TENSORCORE_RESNET_COMMON_FUNCTIONAL_ADD_HPP
