+++ resnet/tests/functional/gemm.cpp
@@ -153,10 +153,10 @@ static void func_test(size_t M,
   default_random_engine generator(RANDOM_SEED);
   uniform_real_distribution<float> matrix_dist(-1.0e2, 1.0e2);
 
-  for (int i = 0; i < M * K; i++) {
+  for (size_t i = 0; i < M * K; i++) {
     A[i] = matrix_dist(generator);
   }
-  for (int i = 0; i < K * N; i++) {
+  for (size_t i = 0; i < K * N; i++) {
     B[i] = matrix_dist(generator);
   }
 
@@ -190,7 +190,7 @@ static void func_test(size_t M,
 
   // Check if C is correct
   double average_diff = 0;
-  for (int i = 0; i < M * N; i++) {
+  for (size_t i = 0; i < M * N; i++) {
     auto diff_ratio = C_ref[i] != 0 ? fabs(C[i] - C_ref[i]) / fabs(C_ref[i]) : 0;
     average_diff += diff_ratio / (float) (M * N);
   }
@@ -221,13 +221,13 @@ static void func_test(size_t M,
   default_random_engine generator(RANDOM_SEED);
   uniform_real_distribution<float> matrix_dist(-1.0e2, 1.0e2);
 
-  for (int i = 0; i < M * K; i++) {
+  for (size_t i = 0; i < M * K; i++) {
     // Make sure the matrix have the same float
     float_16 a_float_16 = __float2half(matrix_dist(generator));
     A[i] = __half2float(a_float_16);
     A_float_16[i] = a_float_16;
   }
-  for (int i = 0; i < K * N; i++) {
+  for (size_t i = 0; i < K * N; i++) {
     float_16 b_float_16 = __float2half(matrix_dist(generator));
     B[i] = __half2float(b_float_16);
     B_float_16[i] = b_float_16;
@@ -284,7 +284,7 @@ static void func_test(size_t M,
 
   // Check if C is correct
   double average_diff = 0;
-  for (int i = 0; i < M * N; i++) {
+  for (size_t i = 0; i < M * N; i++) {
     auto diff_ratio = C_ref[i] != 0 ? fabs(C[i] - C_ref[i]) / fabs(C_ref[i]) : 0;
     average_diff += diff_ratio / (float) (M * N);
   }
